Rename gcsbucket to storage
